+++ src/features/linodes/LinodesDetail/LinodeAdvanced/index.ts
@@ -1,0 +1,2 @@
+import LinodeAdvancedConfigurationsPanel from './LinodeAdvancedConfigurationsPanel';
+export default LinodeAdvancedConfigurationsPanel;
